requirements.txt
twilio==8.12.0
openai==1.12.0
python-dotenv==1.0.1
schedule==1.2.1
requests==2.31.0
loguru==0.7.2
